feat: extract community feed filter logic to commons
Move CommunityFeedFilter and CommunityModerationFeedFilter logic to
commons module. The filtering, sorting, and matching logic is now in
commons open classes. Android subclasses override feed() to provide
LocalCache iteration.

Part of KMP iOS migration tracked in #2238.
